low-poly BT blocks, new materials, textures, scene rebuilt
The high-poly blocks that used 4 cardinal faces and the bevels between them were removed from the Booster Tower scene. The upper-portion assets used in the scene were all moved down 1 unit.

The blocks were replaced by a low-poly proxy for all the blocks in the scene. I baked textures for this asset in Blender and used them in Unity to make a new, pleasing material. This includes basecolor, normal, and AO. Each of these textures is 2K but costs less than 0.5mb. We may be able to shrink these textures to 1K (especially the basecolor) without loosing too much detail.

The lightmaps were redone. The scene looks extremely similar to how it did before, but the lighting artifact where one of the panels was conspicuously brighter than the others, is gone now. Big improvement.  The scene looks great!

Issues:

* Noticed that the item from the ItemUse (sic) action is not actually consumed; it is available on the next turn. Ask Victor if this is for debugging purposes.

* Item still appears slightly too late, maybe 0.1-0.2 seconds too late.

TODO:
* Make a low-poly (plane) asset for the floor in the Battle Arena scene.
* Discard block FBX and Prefab assets.
* Fireball shader

FOR THE FUTURE:

The colors used by the blocks in Booster's Tower are:
D99A41 (yellow)
0A3740 (blue)