"Falling" animation for Luigi_Jump_Down_Beginning state
Wow, what a ride! First, VS2019 said the project was incompatable. The old problem we used to have. Fixing that took a day of moaning and an hour of head-scratching (don't tell nobody though!). The solution to that was to actually check the Visual Studio extensions and make sure the Unity addon is enabled--probably. I had to reenable it but I also deleted and regenerated the VS project files in the root, just to be safe.

After that, I spent a while tinkering with the Jump attack timing and logic. I "fixed it" until I was able to add an sfx event without trouble. Speaking of which: The falling sound effect is from freesfx.com. We just need to credit them before our demo goes out into the wild. The sfx is in the game right now. We might not want it to play on repeated jumps once the timed hits for that are implemented.