culled test scenes; updates to AudioManager.cs, SFX_Events.cs
I wanted to remove more unused assets but didn't want to get us in trouble.
I left the "PersistentScene" alone (even though we are not using it yet...I hope), only deleting the "test" and "thwompTest" scenes. Other changes:

Added and using "AudioManager" tag;
Changed ExitTime, Duration of transition between the two animator trees;
Appended a transition that takes us from the Battle tree to the Platforming tree. Not currently in use.
Altered AudioManager.cs and SFX_Events.cs (see below).

I wanted to improve debugging for sfx, and got mixed results. Before, a LogError was returned if SetSFX() was called and it couldn't match the animation names to cases. SOMETHING was going wrong because those logs were coming when I didn't expect them. So I changed the logic to report an error if animation name isn't in a list stored on the AudioManger.cs script. This didn't fix anything.

What's weird is that SetSFX() is (supposedly) getting called by the animation whose name is "Luigi|Jump_Down", but that animation is read-only, meaning I wasn't able to set an animation event for it that would call the script in the first place. That animation is NOT SUPPOSED TO HAVE SOUND! We shouldn't get an error log because the script shouldn't run. You with me?

Ok, so what might be relevant is that I had to CTRL+D (duplicate) every animation I wanted an animation event for, rename it, and those animations' events call this script. One of the animations I copied IS "Luigi|Jump_Down", but I renamed it and it isn't used in the tree that is accessed by the animator when the BoosterTower scene is running. I get this error in that scene, only some of the time.

So the question is: When Luigi is in the Jump_Down state in the Platforming tree, is there some kind of metadata that is mistakenly pointing to the duplicate animation that isn't being played, yet reporting the name associated with the animation that is being played--SOME OF THE TIME?

It is possible that all I have done is over-complicate a simple task. Undoing the change is easy if we can't build upon it for something useful.