Fixed Thwomp's normals
Very confused. The metadata says it is setting blendShapeNormalImportMode to 0, but I would imagine the fix had me setting it to 1? Guess it's backwards. That would explain the bugs people used to complain about. It's self-explanatory in the editor. You go to Mesh -> BlendShape Normals and choose "Imported". After that, even if your choose "Calculate" again (the default), the normals won't be messed up anymore.