changes to AI_Thwomp.cs and AnimCntrlr_Thwomp42ef.controller (Mechanim)
Thwomp's Mechanim:
- changed play speed of the reverse-grimmace to -0.333;
- changed various transition parameters that were causing some issues

AI_Thwomp.cs:
- renamed "Thwomp" action states;
- changed some timings;

NOTICED A BUG:

Way back when, I wrote a function called "LerpOverTime()". I thought this was a no-brainer. However, I have noticed that if you choose any duration input over 1.0f, the result is not as expected. I noticed that a 10.0 second lerp doesn't take nearly that long and eases-out to its finish position. This is especially noticeable if you crank the duration up to 1000f. Further, it seems that regardless of the value inputed, the end position seems to be reached early. This might have something to do with the division? I am doing float/float, I thought, so there should be no issue. I also track the percent value with Debug Logs and those seem to work fine too. Yet Thwomp is reaching his endpoint in roughly half the time, and the transition is smooth, not linear. WTF??? Is this function being called more than once per frame? What is with the smoothing at the end? Needs to be an issue.