Added "ChimpInTheBox" asset to the project.
"Chimp-In-The-Box" (not final name) is in the project. He has 7 animations and a Ref-pose. It took some-doing to get the animations to work in Unity. Because of this, I am much more aware of animation compatibility issues between Blender and Unity. There is no tutorial for what I have done. I should make one.

Making this character also enabled me to update the ItemBox material. We do not need to distinguish between ItemBoxWood and ItemBoxMetal. However, the asset itself still has 2 material slots. Whatever.

This character probably (not definitely) requires no transitions because every one of his bones is disconnected and I do not know how well they will be interpolated by Unity if we let it handle transitions between animation states. We can see, of course.

This character will "need" the question mark programmed to turn off when the box closes, similarly to how we have done for the ItemBox asset.